begin RenderGraph module
Currently only linearizes all passes and submits to single queue,
lots of potential for GPU saturation here. The most important
benefit right now is that each graphics algorithm can be defined
in isolation and with encapsulation. The render graph manages
pass dependencies and image layout transitions for all images
declared with the API.